+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
  - Only updating relevant displays presenters halves routing feedback time
 
+## [9.1.0] - 2019-08-27
+### Added
+ - CUE background mode controlled by ConnectPRO with CueBackground config element on ConnectProTheme
+ - Multiple camera support in Zoom and other wtc/vtc.
+ - Added camera active and camera control presenters.
+ - Added camera active selection and camera control subpage views.
+
 ## [9.0.0] - 2019-08-15
 ### Added
  - Added CUE Conference page to show instead of routing the Zoom Room until the device is in a meeting